% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplots.R
\name{add_barplot}
\alias{add_barplot}
\title{Add a ggplotly bar plot to a chronicle report}
\usage{
add_barplot(
  report = new_report(),
  dt,
  bars,
  value = NULL,
  break_bars_by = NULL,
  ggtheme = NULL,
  barplot_title = NULL,
  title_level = 2,
  x_axis_label = NULL,
  y_axis_label = NULL,
  plot_palette = c("#58508d", "#ffa600", "#ff6361", "#003f5c", "#bc5090", "#A6CEE3",
    "#1F78B4", "#B2DF8A", "#33A02C", "#FB9A99", "#E31A1C", "#FDBF6F", "#FF7F00",
    "#CAB2D6", "#6A3D9A", "#FFFF99")
)
}
\arguments{
\item{report}{character containing the text of an Rmarkdown report header (and possibly more chunks). Easily create one with chronicle::new_report()}

\item{dt}{data.frame containing the data to plot. It must have a numerical variable, a date variable, and optionally a grouping variable to split the data and plot them as individual time sieries inside the same plot.}

\item{bars}{Name of the columns containing the different groups.}

\item{value}{Name of the columns to use as values on the y axis of the plot. If NULL, counts will be used.}

\item{break_bars_by}{Name of the categorical variable used to break each bar}

\item{ggtheme}{ggplot2 theme function to apply. Default is ggplot2::theme_minimal.}

\item{barplot_title}{Title of the dygraph.}

\item{title_level}{Level of the section title of this plot (ie, number of # on Rmarkdown syntax.)}

\item{x_axis_label}{Label for the x axis.}

\item{y_axis_label}{Label for the y axis.}

\item{plot_palette}{Character vector of hex codes specifying the colors to use on the plot. Default is RColorBrewer's Paired and Spectral colors concatenated.}
}
\value{
The text of the Rmarkdown report plus an additional section with the bar plot.
}
\description{
Add a ggplotly bar plot to a chronicle report
}
