% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{add_boxplot}
\alias{add_boxplot}
\title{Add a ggplotly boxplot to a chronicle report}
\usage{
add_boxplot(
  report = new_report(),
  dt,
  num_var,
  group_var = NULL,
  ggtheme = NULL,
  boxplot_title = NULL,
  title_level = 2,
  x_axis_label = NULL,
  y_axis_label = NULL,
  y_axis_limits = NULL,
  plot_palette = c("#58508d", "#ffa600", "#ff6361", "#003f5c", "#bc5090", "#A6CEE3",
    "#1F78B4", "#B2DF8A", "#33A02C", "#FB9A99", "#E31A1C", "#FDBF6F", "#FF7F00",
    "#CAB2D6", "#6A3D9A", "#FFFF99")
)
}
\arguments{
\item{report}{character containing the text of an Rmarkdown report header (and possibly more chunks). Easily create one with chronicle::new_report()}

\item{dt}{data.frame containing the data to plot. It must have a numerical variable, a date variable, and optionally a grouping variable to split the data and plot them as individual time sieries inside the same plot.}

\item{num_var}{Name of the column of the data frame containing the numerical variable.}

\item{group_var}{Name of the columns containing the different groups.}

\item{ggtheme}{ggplot2 theme function to apply. Default is ggplot2::theme_minimal.}

\item{boxplot_title}{Title of the dygraph.}

\item{title_level}{Level of the section title of this plot (ie, number of # on Rmarkdown syntax.)}

\item{x_axis_label}{Label for the x axis.}

\item{y_axis_label}{Label for the y axis.}

\item{y_axis_limits}{Numeric vector of length 2 specifying the lower and upper bounds of the y axis (useful when outliers eclipse the interquartile range).}
}
\value{

}
\description{
Add a ggplotly boxplot to a chronicle report
}
