% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{add_title}
\alias{add_title}
\title{Add a titled section to a chronicle Rmarkdown report}
\usage{
add_title(report, title, title_level = 1)
}
\arguments{
\item{report}{character containing the text of an Rmarkdown report header (and possibly more chunks). Easily create one with chronicle::new_report()}

\item{title}{The title to be added as a section.}

\item{title_level}{Level of the section title (ie, number of # on Rmarkdown syntax.)}
}
\value{
The text of the Rmarkdown report plus an additional section with the text
}
\description{
Add a titled section to a chronicle Rmarkdown report
}
