% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_interface.R
\name{file_extension}
\alias{file_extension}
\title{Parse the file extension for each R Markdown output format}
\usage{
file_extension(file_type)
}
\arguments{
\item{file_type}{R Markdown output formats.}
}
\value{
The file extension corresponding to the provided formats (".html", "pdf", ".md", ".docx", ".pptx")
}
\description{
Currently supports:
}
\details{
*  rmdformats
 *  prettydoc
 *  bookdown
 *  ioslides
 *  tufte_html
 *  xaringan
 *  rolldown
 *  flexdashboard
 *  slidy_presentation
 *  html_document
 *  html_notebook
 *  pagedown
}
\examples{
file_extension(c('prettydoc', 'word_document', 'tufte_handout'))
}
