% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enchunk.R
\name{make_title}
\alias{make_title}
\title{Guess a title out of function parameters}
\usage{
make_title(fun, params)
}
\arguments{
\item{fun}{chronicle make_* function}

\item{params}{parameters for fun}
}
\value{
A generic title for the plot
}
\description{
Detects which make_* function is passed and builds a generic name based on its parameters.
}
\examples{
make_title(fun = chronicle::make_barplot,
                     params = list(value = 'Amount',
                                   bars = 'Country',
                                   break_bars_by = 'Region'))

make_title(fun = chronicle::make_raincloud,
           params = list(value = 'value',
                         groups = 'species'))
}
