% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_interface.R
\name{new_report}
\alias{new_report}
\title{Create the initial Rmarkdown header for a report}
\usage{
new_report(
  title = "New chronicle Report",
  author = "chronicle user",
  prettydoc = TRUE,
  prettydoc_theme = "leonids",
  highlight = "github",
  number_sections = FALSE,
  table_of_content = FALSE,
  table_of_content_depth = 1
)
}
\arguments{
\item{title}{Title of the report.}

\item{author}{Author of the repor.}

\item{prettydoc}{Logical indicating whether or not to use prettydoc formatting on the html report.}

\item{prettydoc_theme}{Name of the theme used on prettydoc. Default is cayman.}

\item{highlight}{Rmarkdown highlight theming}

\item{number_sections}{Whether or not to numnber the sections and subsections fo the report.}

\item{table_of_content}{Whether or not to build a table fo content at the begining of the report.}

\item{table_of_content_depth}{The depth of sections and subsections to be displayed on the table of content.}
}
\value{
a String conainitng an Rmarkdown header
}
\description{
Create the initial Rmarkdown header for a report
}
\examples{
new_report('Simple Report Header', author = 'Anonymous Developer', prettydoc = FALSE)
new_report('Prettier Report Header', author = 'The same person as before', prettydoc = TRUE)
}
