% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_columns.R
\name{plot_columns}
\alias{plot_columns}
\title{Plot all columns of a table}
\usage{
plot_columns(dt, by_column = NULL)
}
\arguments{
\item{dt}{Table to be plotted.}

\item{by_column}{Name of the colum to use as groups for all the other plots}
}
\value{
A list of plotly::ggplotly objects, one for each column of the table.
}
\description{
Make boxplots for each numerical variable on a table, and barplots for each categorical variable.
}
\examples{
chronicle::report_columns(dt = iris, by_column = 'Species')
}
