% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enchunk.R
\name{rmd_title_level}
\alias{rmd_title_level}
\title{Returns the count of '#' corresponding to a given title level}
\usage{
rmd_title_level(level)
}
\arguments{
\item{level}{R Markdonw title level}
}
\value{
'#', '##', '###' and so on, depending on the title level
}
\description{
Returns the count of '#' corresponding to a given title level
}
\examples{
rmd_title_level(1)
rmd_title_level(3)
}
