% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{set_classes}
\alias{set_classes}
\title{Change column classes with a named vector}
\usage{
set_classes(
  dt,
  character = NULL,
  integer = NULL,
  double = NULL,
  logical = NULL,
  factor = NULL
)
}
\arguments{
\item{dt}{Table whose column types will be changed by reference}

\item{character}{A character vector specifiying which columns that will be coerced to character.}

\item{integer}{A character vector specifiying which columns that will be coerced to integer.}

\item{double}{A character vector specifiying which columns that will be coerced to double.}

\item{logical}{A character vector specifiying which columns that will be coerced to logical.}

\item{factor}{A character vector specifiying which columns that will be coerced to factor.}
}
\value{
Changes by reference the types of the specified columns
}
\description{
Change column classes with a named vector
}
\examples{
iris_changed <- set_classes(dt = idt, character = 'Species', integer = c('Sepal.Length', 'Sepal.Width'))
purrr::map(iris_changed, class)
}
