% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_interface.R
\name{update_chronicle}
\alias{update_chronicle}
\title{Wrapper to update chronicle from github}
\usage{
update_chronicle()
}
\description{
Quality of life improvement until I submit the package to CRAN. Runs devtools::install_github('pheymanss/chronicle')
}
