% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{add_code}
\alias{add_code}
\title{#' Add formated code chunks to a chronicle Rmarkdown report}
\usage{
add_code(report, code, code_title = NULL, title_level = 2)
}
\arguments{
\item{report}{character containing the text of an Rmarkdown report header (and possibly more chunks). Easily create one with chronicle::new_report()}

\item{code}{Character string to be formated as an R code chunk. This code will not be excecuted when the report is rendered.}

\item{code_title}{The title of the code section. Default is NULL.}

\item{title_level}{Level of the section title (ie, number of # on Rmarkdown syntax.)}

\item{title}{The title to be added as a section.}
}
\value{
The text of the Rmarkdown report plus an additional section with the code chunk.
}
\description{
#' Add formated code chunks to a chronicle Rmarkdown report
}
