% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplots.R
\name{make_barplot}
\alias{make_barplot}
\title{create a bar plot from a data frame through ggplotly}
\usage{
make_barplot(
  dt,
  bars,
  value,
  break_bars_by = NULL,
  ggtheme = ggplot2::theme_minimal,
  x_axis_label = NULL,
  y_axis_label = NULL,
  plot_palette = c("#58508d", "#ffa600", "#ff6361", "#003f5c", "#bc5090", "#A6CEE3",
    "#1F78B4", "#B2DF8A", "#33A02C", "#FB9A99", "#E31A1C", "#FDBF6F", "#FF7F00",
    "#CAB2D6", "#6A3D9A", "#FFFF99")
)
}
\arguments{
\item{dt}{data.frame containing the data to plot. It must have a numerical variable, a date variable, and optionally a grouping variable to split the data and plot them as individual time sieries inside the same plot.}

\item{bars}{Name of the columns containing the different groups.}

\item{value}{Name of the columns to use as values on the y axis of the plot. If NULL, counts will be used.}

\item{break_bars_by}{Name of the categorical variable used to break each bar}

\item{ggtheme}{ggplot2 theme function to apply. Default is ggplot2::theme_minimal.}

\item{x_axis_label}{Label for the x axis.}

\item{y_axis_label}{Label for the y axis.}

\item{plot_palette}{#' @param plot_palette Character vector of hex codes specifying the colors to use on the plot.}
}
\value{
a plotly-ized version of a grouped ggplot bar plot
}
\description{
create a bar plot from a data frame through ggplotly
}
