% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_interface.R
\name{output_config}
\alias{output_config}
\title{Wrapper to update chronicle from github}
\usage{
output_config(
  output_format,
  number_sections = FALSE,
  table_of_content = FALSE,
  table_of_content_depth = 1,
  fig_width = 11,
  fig_height = 5,
  prettydoc_theme = "leonids"
)
}
\arguments{
\item{output_format}{The format of the R Markdown file.}

\item{number_sections}{Whether or not to number the sections and subsections of the report.}

\item{table_of_content}{Whether or not to include a table fo content at the beginning of the report.}

\item{table_of_content_depth}{The depth of sections and subsections to be displayed on the table of content.}

\item{fig_width}{Set the global figure width or the rmarkdown file.}

\item{fig_height}{Set the global figure height or the rmarkdown file.}

\item{prettydoc_theme}{Name of the theme used on prettydoc. Default is leonids.}
}
\value{
The lines needed in the yaml header of an R Markdown file to render as the specified output type.
}
\description{
Quality of life improvement until I submit the package to CRAN. Runs devtools::install_github('pheymanss/chronicle')
Build the yaml output specification for an R Markdown
}
\details{
Currently supported: prettydoc, ioslides, tufte, flexdashboard, slidy_presentation,
html_document, html_notebook.
}
\examples{
cat(output_config('prettydoc'))
cat(output_config('ioslides'))
}
